refactor(server): update docs and cleanup validators and constants
