+++ packages/amplication-server/src/core/prismaSchemaUtils/constants.ts
@@ -21,3 +21,16 @@ export const ARG_KEY_FIELD_NAME = "fields";
 export const ID_TYPE_CUID = "CUID";
 export const ID_TYPE_UUID = "UUID";
 export const ID_TYPE_AUTOINCREMENT = "AUTO_INCREMENT";
+
+export const idTypePropertyMap = {
+  autoincrement: ID_TYPE_AUTOINCREMENT,
+  cuid: ID_TYPE_CUID,
+  uuid: ID_TYPE_UUID,
+};
+
+export const idTypePropertyMapByFieldType = {
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  Int: ID_TYPE_AUTOINCREMENT,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  String: ID_TYPE_CUID,
+};
